x/cdac/types/messages_domain_cooperation_test.go
package types

import (
	"testing"

	"crossdomain/testutil/sample"
	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/require"
)

func TestMsgCreateDomainCooperation_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgCreateDomainCooperation
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgCreateDomainCooperation{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgCreateDomainCooperation{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}

func TestMsgUpdateDomainCooperation_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgUpdateDomainCooperation
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgUpdateDomainCooperation{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgUpdateDomainCooperation{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}

func TestMsgDeleteDomainCooperation_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgDeleteDomainCooperation
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgDeleteDomainCooperation{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgDeleteDomainCooperation{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}
